Fix MSVC /analyze C4244 in window-mode tile coordinate lambda
std::pair<uint8_t,uint8_t> was being constructed from {m_tileX % 32,
m_Y % 256}, both int-promoted expressions - a narrowing conversion
that happens inside std::pair's converting constructor rather than
at the call site, which GCC/Clang's -Wconversion doesn't surface
(diagnostics originating inside standard-library code are
suppressed by default) but MSVC's C4244 does, under -Werror. Make
the narrowing explicit with static_cast, matching the sibling
non-window branch's existing style.

+++ src/ppu.cpp
@@ -108,7 +108,8 @@ Ppu::Fetcher::runNextTCycle()
         const auto& [tileX,
                      tileY] = [&]() -> std::pair<std::uint8_t, std::uint8_t> {
           if (m_mode == Mode::Window) {
-            return { m_tileX % 32, m_Y % 256 };
+            return { static_cast<std::uint8_t>(m_tileX % 32),
+                     static_cast<std::uint8_t>(m_Y % 256) };
           }
           const std::uint8_t tileXPrim =
             (m_tileX + (m_mmu.get().readByte(regs::SCX) / 8)) % 32;
